added error handling to the change password

See merge request ayush.singhal/notifications-microservice!27

+++ api/serializers/get_recipient.go
@@ -6,7 +6,7 @@ import (
 
 // RecipientInfo serializer to bind request data
 type RecipientInfo struct {
-	ID                 uint   `json:"recipient_id"`
+	ID                 uint64 `json:"recipient_id"`
 	Email              string `json:"email,omitempty"`
 	PushToken          string `json:"push_token,omitempty"`
 	WebToken           string `json:"web_token,omitempty"`
@@ -15,7 +15,7 @@ type RecipientInfo struct {
 
 // RecipientModelToRecipientInfo converts the Recipient model to RecipientInfo struct
 func RecipientModelToRecipientInfo(info *RecipientInfo, recipient *models.Recipient) {
-	info.ID = recipient.ID
+	info.ID = uint64(recipient.ID)
 	info.Email = recipient.Email
 	info.PushToken = recipient.PushToken
 	info.WebToken = recipient.WebToken
